requirements.txt
environs==9.5.0
python-telegram-bot==13.0
redis==4.5.1
requests==2.28.2